feat: folder structure refsctoring
