+++ cdk/services/dynamodb/getProductsTable.ts
@@ -3,54 +3,54 @@ import * as cdk from "aws-cdk-lib";
 import { ITable} from "aws-cdk-lib/aws-dynamodb";
 import * as dynamodb from 'aws-cdk-lib/aws-dynamodb'
 //import { DynamoDB, } from '@aws-sdk/client-dynamodb';
-import { DYNAMODB } from '../../../constants/constants';
+import { DYNAMODB } from '../../constants/constants';
 
 
-export class GetStocksTable {
-    getStocksTable: ITable
+export class GetProductsTable {
+    getProductsTable: ITable
 
     constructor(scope: Construct, id: string) {
 
-        const getStocksTable = new dynamodb.Table(
+        const getProductsTable = new dynamodb.Table(
             scope,
             id,
             {
-              partitionKey: { name: "product_id", type: dynamodb.AttributeType.STRING },
-              tableName: DYNAMODB.STOCKS_TABLE,
+              partitionKey: { name: "id", type: dynamodb.AttributeType.STRING },
+              tableName: DYNAMODB.PRODUCTS_TABLE,
               removalPolicy: cdk.RemovalPolicy.DESTROY,
               //removalPolicy: cdk.RemovalPolicy.RETAIN,
             }
         );
 
-        this.getStocksTable = getStocksTable;
+        this.getProductsTable = getProductsTable;
 
 /*
         const dynamoDB = new DynamoDB({});
 
         const tables = dynamoDB.listTables()
 
-        if (tables!.TableNames!.indexOf(DYNAMODB.STOCKS_TABLE) < 0) {
-            const getStocksTable = new dynamodb.Table(
+        if (tables!.TableNames!.indexOf(DYNAMODB.PRODUCTS_TABLE) < 0) {
+            const getProductsTable = new dynamodb.Table(
                 scope,
                 id,
                 {
                   partitionKey: { name: "id", type: dynamodb.AttributeType.STRING },
-                  tableName: DYNAMODB.STOCKS_TABLE,
+                  tableName: DYNAMODB.PRODUCTS_TABLE,
                   removalPolicy: cdk.RemovalPolicy.DESTROY,
                   //removalPolicy: cdk.RemovalPolicy.RETAIN,
                 }
             );
 
-            this.getStocksTable = getStocksTable;
+            this.getProductsTable = getProductsTable;
 
         } else {
-            const getStocksTable = dynamodb.Table.fromTableName(
+            const getProductsTable = dynamodb.Table.fromTableName(
                 scope,
                 id,
-                DYNAMODB.STOCKS_TABLE
+                DYNAMODB.PRODUCTS_TABLE
             )
 
-            this.getStocksTable = getStocksTable;
+            this.getProductsTable = getProductsTable;
         }
 */        
     }
