+++ cdk/services/dynamodb/getStocksTable.ts
@@ -3,54 +3,54 @@ import * as cdk from "aws-cdk-lib";
 import { ITable} from "aws-cdk-lib/aws-dynamodb";
 import * as dynamodb from 'aws-cdk-lib/aws-dynamodb'
 //import { DynamoDB, } from '@aws-sdk/client-dynamodb';
-import { DYNAMODB } from '../../../constants/constants';
+import { DYNAMODB } from '../../constants/constants';
 
 
-export class GetProductsTable {
-    getProductsTable: ITable
+export class GetStocksTable {
+    getStocksTable: ITable
 
     constructor(scope: Construct, id: string) {
 
-        const getProductsTable = new dynamodb.Table(
+        const getStocksTable = new dynamodb.Table(
             scope,
             id,
             {
-              partitionKey: { name: "id", type: dynamodb.AttributeType.STRING },
-              tableName: DYNAMODB.PRODUCTS_TABLE,
+              partitionKey: { name: "product_id", type: dynamodb.AttributeType.STRING },
+              tableName: DYNAMODB.STOCKS_TABLE,
               removalPolicy: cdk.RemovalPolicy.DESTROY,
               //removalPolicy: cdk.RemovalPolicy.RETAIN,
             }
         );
 
-        this.getProductsTable = getProductsTable;
+        this.getStocksTable = getStocksTable;
 
 /*
         const dynamoDB = new DynamoDB({});
 
         const tables = dynamoDB.listTables()
 
-        if (tables!.TableNames!.indexOf(DYNAMODB.PRODUCTS_TABLE) < 0) {
-            const getProductsTable = new dynamodb.Table(
+        if (tables!.TableNames!.indexOf(DYNAMODB.STOCKS_TABLE) < 0) {
+            const getStocksTable = new dynamodb.Table(
                 scope,
                 id,
                 {
                   partitionKey: { name: "id", type: dynamodb.AttributeType.STRING },
-                  tableName: DYNAMODB.PRODUCTS_TABLE,
+                  tableName: DYNAMODB.STOCKS_TABLE,
                   removalPolicy: cdk.RemovalPolicy.DESTROY,
                   //removalPolicy: cdk.RemovalPolicy.RETAIN,
                 }
             );
 
-            this.getProductsTable = getProductsTable;
+            this.getStocksTable = getStocksTable;
 
         } else {
-            const getProductsTable = dynamodb.Table.fromTableName(
+            const getStocksTable = dynamodb.Table.fromTableName(
                 scope,
                 id,
-                DYNAMODB.PRODUCTS_TABLE
+                DYNAMODB.STOCKS_TABLE
             )
 
-            this.getProductsTable = getProductsTable;
+            this.getStocksTable = getStocksTable;
         }
 */        
     }
