+++ cdk/services/import-service/ImportService.ts
@@ -2,9 +2,9 @@ import { Construct } from 'constructs';
 //import { ImportBucket } from '../../s3/import/importBucket';
 import { ImportProductsFile } from './importProductsFile';
 import { ApiGatewayImportService } from './ApiGatewayImportService';
-import { QUEUE, S3 } from '../../../constants/constants';
+import { QUEUE, S3 } from '../../constants/constants';
 import { ImportFileParser } from './importFileParser'; 
-import { taskNum } from '../../../constants/constants';
+import { taskNum } from '../../constants/constants';
 import { Bucket, EventType } from 'aws-cdk-lib/aws-s3';
 import { LambdaDestination } from "aws-cdk-lib/aws-s3-notifications";
 //import { S3EventSource } from 'aws-cdk-lib/aws-lambda-event-sources';
