+++ cdk/services/product-service/ProductService.ts
@@ -3,12 +3,12 @@ import { GetProductsList } from './GetProductsList';
 import { GetProductById } from './GetProductsById';
 import { CreateProduct } from './CreateProduct';
 import { ApiGatewayProductService } from './ApiGatewayProductService';
-import { GetProductsTable } from '../../dynamodb/product-storage/getProductsTable';
-import { GetStocksTable } from '../../dynamodb/product-storage/getStocksTable';
-import { taskNum } from '../../../constants/constants';
+import { GetProductsTable } from '../dynamodb/getProductsTable';
+import { GetStocksTable } from '../dynamodb/getStocksTable';
+import { taskNum } from '../../constants/constants';
 import { Queue } from 'aws-cdk-lib/aws-sqs';
 import { CreateProductTopic } from './CreateProductTopic';
-import { QUEUE } from '../../../constants/constants';
+import { QUEUE } from '../../constants/constants';
 import { CatalogBatchProcessLambda } from './catalogBatchProcess';
 import { SqsEventSource } from 'aws-cdk-lib/aws-lambda-event-sources';
 import { CfnOutput } from 'aws-cdk-lib';
