+++ cdk/services/product-service/catalogBatchProcess.ts
