+++ cdk/services/product-service/functions/catalogBatchProcessLambda.test.ts
