+++ cdk/services/product-service/functions/cdk-products.ts
@@ -1,5 +1,5 @@
     import { APIGatewayProxyEvent, APIGatewayProxyHandler, APIGatewayProxyResult } from 'aws-lambda';
-    import { getProductsData } from '../../../dynamodb/product-storage/getProductsData';
+    import { getProductsData } from '../../dynamodb/getProductsData';
 
 
     export const handler: APIGatewayProxyHandler = async (
