+++ cdk/services/product-service/snsClient.ts
